+++ CHANGELOG.md
@@ -2,5 +2,10 @@
 Changelog for Shipflo WordPress Plugin
 
 ## 1.0.0.1 2025-10-30
-RELEASE: Initial stable release 
+[RELEASE]: Initial stable release 
 
+
+## 1.1.0.0 2025-11-04
+[UPDATE]: Updated client side logic
+[UPDATE]: Updated client side key encryption 
+[UPDATE]: Updated client order filteration logic